+++ test/unit/dialogs.spec.js
@@ -1,7 +1,13 @@
 import createDialogs from '../../lib/api/dialogs'
 import { describeChannelCallingMethod } from '../helpers'
 
-const methods = [
+const SIMPLE_DIALOGS = [
+  ['openAlert', 'alert'],
+  ['openConfirm', 'confirm'],
+  ['openPrompt', 'prompt']
+]
+
+const ENTITY_SELECTOR_DIALOGS = [
   ['selectSingleEntry', 'Entry', false],
   ['selectSingleAsset', 'Asset', false],
   ['selectMultipleEntries', 'Entry', true],
@@ -10,14 +16,23 @@ const methods = [
 
 describe('createDialogs()', () => {
   describe('returned "dialogs" object', () => {
-    methods.forEach(([method, entityType, multiple]) => {
-      const callOptions = {test: true, entityType, multiple}
+    SIMPLE_DIALOGS.forEach(([method, type]) => {
       describeChannelCallingMethod({
         creator: createDialogs,
         methodName: method,
         channelMethod: 'openDialog',
         args: [{test: true}],
-        expectedCallArgs: ['entitySelector', callOptions]
+        expectedCallArgs: [type, {test: true}]
+      })
+    })
+
+    ENTITY_SELECTOR_DIALOGS.forEach(([method, entityType, multiple]) => {
+      describeChannelCallingMethod({
+        creator: createDialogs,
+        methodName: method,
+        channelMethod: 'openDialog',
+        args: [{test: true}],
+        expectedCallArgs: ['entitySelector', {test: true, entityType, multiple}]
       })
     })
   })